Added save evaluation validator
