+++ src/main/java/br/com/annuum/capsicum/api/controller/SkillController.java
@@ -1,7 +1,6 @@
 package br.com.annuum.capsicum.api.controller;
 
 import br.com.annuum.capsicum.api.controller.response.SkillListResponse;
-import br.com.annuum.capsicum.api.domain.dto.SkillDto;
 import br.com.annuum.capsicum.api.service.FindAllSkillsService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -9,8 +8,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("/skill")
