+++ src/main/java/br/com/annuum/capsicum/api/service/SaveVolunteerEvaluationService.java
@@ -4,10 +4,9 @@ import br.com.annuum.capsicum.api.controller.request.VolunteerEvaluationRequest;
 import br.com.annuum.capsicum.api.domain.Candidacy;
 import br.com.annuum.capsicum.api.domain.Movement;
 import br.com.annuum.capsicum.api.domain.VolunteerEvaluation;
-import br.com.annuum.capsicum.api.domain.enums.CandidacyStatus;
-import br.com.annuum.capsicum.api.domain.enums.MovementStatus;
 import br.com.annuum.capsicum.api.exceptions.AccessControlException;
 import br.com.annuum.capsicum.api.repository.VolunteerEvaluationRepository;
+import br.com.annuum.capsicum.api.validator.SaveEvaluationValidator;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -27,6 +26,9 @@ public class SaveVolunteerEvaluationService {
     @Autowired
     private FindCandidacyByIdService findCandidacyByIdService;
 
+    @Autowired
+    private SaveEvaluationValidator saveEvaluationValidator;
+
     @Transactional
     public VolunteerEvaluation save(final Long idUserAuthenticated, final VolunteerEvaluationRequest volunteerEvaluationRequest) {
 
@@ -39,13 +41,7 @@ public class SaveVolunteerEvaluationService {
             throw new AccessControlException("O usuário autenticado não é o autor do movimento.");
         }
 
-        if (!movement.getMovementStatus().equals(MovementStatus.CONCLUDE)) {
-            throw new AccessControlException("Não é possível avaliar voluntários enquanto o Movimento não estiver concluído.");
-        }
-
-        if (!candidacy.getCandidacyStatusControl().getStatusEnum().equals(CandidacyStatus.PRESENT)) {
-            throw new AccessControlException("Não é possível avaliar um voluntário com status da candidatura diferente de PRESENT.");
-        }
+        saveEvaluationValidator.validate(movement, candidacy);
 
         log.info("Building VolunteerEvaluation to persist");
         final VolunteerEvaluation volunteerEvaluation = new VolunteerEvaluation()
