+++ src/main/java/br/com/annuum/capsicum/api/validator/SaveEvaluationValidator.java
@@ -1,0 +1,23 @@
+package br.com.annuum.capsicum.api.validator;
+
+import br.com.annuum.capsicum.api.domain.Candidacy;
+import br.com.annuum.capsicum.api.domain.Movement;
+import br.com.annuum.capsicum.api.domain.enums.CandidacyStatus;
+import br.com.annuum.capsicum.api.domain.enums.MovementStatus;
+import br.com.annuum.capsicum.api.exceptions.AccessControlException;
+import org.springframework.stereotype.Component;
+
+@Component
+public class SaveEvaluationValidator {
+
+    public void validate(Movement movement, Candidacy candidacy) {
+
+        if (!movement.getMovementStatus().equals(MovementStatus.CONCLUDE)) {
+            throw new AccessControlException("Não é possível avaliar voluntários enquanto o Movimento não estiver concluído.");
+        }
+
+        if (!candidacy.getCandidacyStatusControl().getStatusEnum().equals(CandidacyStatus.PRESENT)) {
+            throw new AccessControlException("Não é possível avaliar um voluntário com status da candidatura diferente de PRESENT.");
+        }
+    }
+}
