+++ src/main/resources/application.yml
@@ -1,7 +1,7 @@
 annuum:
   image-repo:
-      url: https://annuum-image-api-homolog.herokuapp.com/
-      token: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJjYXBzaWN1dW0tYW5udXVtIiwiaWF0IjoxNTkzMzcwNTkwLCJleHAiOjQwODYzNTYxOTAsImF1ZCI6IiIsInN1YiI6ImNhcHNpY3V1bS1hbm51dW0tYXBwLWhvbW9sb2cifQ.Z9tU8jf4sgnxTB4cedYwblfcNwzIHVQhR7hE3TLHypg
+    url: https://annuum-image-api-homolog.herokuapp.com/
+    token: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJjYXBzaWN1dW0tYW5udXVtIiwiaWF0IjoxNTkzMzcwNTkwLCJleHAiOjQwODYzNTYxOTAsImF1ZCI6IiIsInN1YiI6ImNhcHNpY3V1bS1hbm51dW0tYXBwLWhvbW9sb2cifQ.Z9tU8jf4sgnxTB4cedYwblfcNwzIHVQhR7hE3TLHypg
   security:
     cors:
       allowed-origins: "*"
