+++ src/test/java/br/com/annuum/capsicum/api/service/SaveMovementAuthorEvaluationServiceTest.java
@@ -6,10 +6,9 @@ import br.com.annuum.capsicum.api.domain.enums.CandidacyStatus;
 import br.com.annuum.capsicum.api.domain.enums.MovementStatus;
 import br.com.annuum.capsicum.api.exceptions.AccessControlException;
 import br.com.annuum.capsicum.api.repository.MovementAuthorEvaluationRepository;
+import br.com.annuum.capsicum.api.validator.SaveEvaluationValidator;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.EnumSource;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Mockito;
@@ -33,6 +32,9 @@ class SaveMovementAuthorEvaluationServiceTest {
 
     @Mock
     private FindCandidacyByIdService findCandidacyByIdService;
+
+    @Mock
+    private SaveEvaluationValidator saveEvaluationValidator;
 
     @Test
     public void mustSaveNewMovementEvaluation_withSuccess() {
@@ -69,6 +71,7 @@ class SaveMovementAuthorEvaluationServiceTest {
             .thenReturn(movement);
         Mockito.when(movementAuthorEvaluationRepository.save(expectedMovementAuthorEvaluation))
             .thenReturn(expectedMovementAuthorEvaluation);
+        Mockito.doNothing().when(saveEvaluationValidator).validate(movement, candidacy);
 
         //Act
         final MovementAuthorEvaluation returnedMovementAuthorEvaluation =
@@ -96,137 +99,6 @@ class SaveMovementAuthorEvaluationServiceTest {
         final Movement movement = new Movement()
             .setUserAuthor(userEvaluated)
             .setMovementStatus(MovementStatus.CONCLUDE);
-
-        final MovementAuthorEvaluationRequest movementAuthorEvaluationRequest = new MovementAuthorEvaluationRequest()
-            .setNote(1)
-            .setFeedback("someFeedBack")
-            .setCandidacyId(1L);
-
-        final Long userEvaluatorId = 2L;
-
-        Mockito.when(findCandidacyByIdService.find(movementAuthorEvaluationRequest.getCandidacyId()))
-            .thenReturn(candidacy);
-        Mockito.when(findMovimentByNeedService.find(candidacy.getNeed()))
-            .thenReturn(movement);
-
-        assertThrows(AccessControlException.class, () -> saveMovementEvaluationService.save(userEvaluatorId, movementAuthorEvaluationRequest));
-    }
-
-    @Test
-    public void shouldReturnExceptionWhenUserEvaluatorHasStatusCandidate() {
-        // Arrange
-        final UserVolunteer userEvaluator = new UserVolunteer();
-        userEvaluator.setId(1L);
-
-        Candidacy candidacy = new Candidacy()
-            .setUserCandidate(userEvaluator)
-            .setCandidacyStatusControl(new CandidacyStatusControl());
-
-        final UserOrganization userEvaluated = Mockito.mock(UserOrganization.class);
-
-        final Movement movement = new Movement()
-            .setUserAuthor(userEvaluated)
-            .setMovementStatus(MovementStatus.CONCLUDE);
-
-        final MovementAuthorEvaluationRequest movementAuthorEvaluationRequest = new MovementAuthorEvaluationRequest()
-            .setNote(1)
-            .setFeedback("someFeedBack")
-            .setCandidacyId(1L);
-
-        final Long userEvaluatorId = 2L;
-
-        Mockito.when(findCandidacyByIdService.find(movementAuthorEvaluationRequest.getCandidacyId()))
-            .thenReturn(candidacy);
-        Mockito.when(findMovimentByNeedService.find(candidacy.getNeed()))
-            .thenReturn(movement);
-
-        assertThrows(AccessControlException.class, () -> saveMovementEvaluationService.save(userEvaluatorId, movementAuthorEvaluationRequest));
-    }
-
-    @ParameterizedTest
-    @EnumSource(value = CandidacyStatus.class, names = {"APPROVED", "DECLINED", "REJECTED"})
-    public void shouldReturnExceptionWhenUserEvaluatorHasStatusApprovedOrDeclinedOrRejected(CandidacyStatus candidacyStatus) {
-        // Arrange
-        final UserVolunteer userEvaluator = new UserVolunteer();
-        userEvaluator.setId(1L);
-
-        Candidacy candidacy = new Candidacy()
-            .setUserCandidate(userEvaluator)
-            .setCandidacyStatusControl(new CandidacyStatusControl());
-        candidacy.getCandidacyStatusControl().setStatusEnum(candidacyStatus);
-
-        final UserOrganization userEvaluated = Mockito.mock(UserOrganization.class);
-
-        final Movement movement = new Movement()
-            .setUserAuthor(userEvaluated)
-            .setMovementStatus(MovementStatus.CONCLUDE);
-
-        final MovementAuthorEvaluationRequest movementAuthorEvaluationRequest = new MovementAuthorEvaluationRequest()
-            .setNote(1)
-            .setFeedback("someFeedBack")
-            .setCandidacyId(1L);
-
-        final Long userEvaluatorId = 2L;
-
-        Mockito.when(findCandidacyByIdService.find(movementAuthorEvaluationRequest.getCandidacyId()))
-            .thenReturn(candidacy);
-        Mockito.when(findMovimentByNeedService.find(candidacy.getNeed()))
-            .thenReturn(movement);
-
-        assertThrows(AccessControlException.class, () -> saveMovementEvaluationService.save(userEvaluatorId, movementAuthorEvaluationRequest));
-    }
-
-    @Test
-    public void shouldReturnExceptionWhenUserEvaluatorHasStatusAbsent() {
-        // Arrange
-        final UserVolunteer userEvaluator = new UserVolunteer();
-        userEvaluator.setId(1L);
-
-        Candidacy candidacy = new Candidacy()
-            .setUserCandidate(userEvaluator)
-            .setCandidacyStatusControl(new CandidacyStatusControl());
-        candidacy.getCandidacyStatusControl().setStatusEnum(CandidacyStatus.APPROVED);
-        candidacy.getCandidacyStatusControl().setStatusEnum(CandidacyStatus.ABSENT);
-
-        final UserOrganization userEvaluated = Mockito.mock(UserOrganization.class);
-
-        final Movement movement = new Movement()
-            .setUserAuthor(userEvaluated)
-            .setMovementStatus(MovementStatus.CONCLUDE);
-
-        final MovementAuthorEvaluationRequest movementAuthorEvaluationRequest = new MovementAuthorEvaluationRequest()
-            .setNote(1)
-            .setFeedback("someFeedBack")
-            .setCandidacyId(1L);
-
-        final Long userEvaluatorId = 2L;
-
-        Mockito.when(findCandidacyByIdService.find(movementAuthorEvaluationRequest.getCandidacyId()))
-            .thenReturn(candidacy);
-        Mockito.when(findMovimentByNeedService.find(candidacy.getNeed()))
-            .thenReturn(movement);
-
-        assertThrows(AccessControlException.class, () -> saveMovementEvaluationService.save(userEvaluatorId, movementAuthorEvaluationRequest));
-    }
-
-    @ParameterizedTest
-    @EnumSource(value = MovementStatus.class, names = {"ACTIVE", "CANCELLED"})
-    public void mustReturnExceptionWhenEvaluatingUserVolunteerMovementStatusDifferentThanConclude(MovementStatus movementStatus) {
-        // Arrange
-        final UserVolunteer userEvaluator = new UserVolunteer();
-        userEvaluator.setId(1L);
-
-        Candidacy candidacy = new Candidacy()
-            .setUserCandidate(userEvaluator)
-            .setCandidacyStatusControl(new CandidacyStatusControl());
-        candidacy.getCandidacyStatusControl().setStatusEnum(CandidacyStatus.APPROVED);
-        candidacy.getCandidacyStatusControl().setStatusEnum(CandidacyStatus.PRESENT);
-
-        final UserOrganization userEvaluated = Mockito.mock(UserOrganization.class);
-
-        final Movement movement = new Movement()
-            .setUserAuthor(userEvaluated)
-            .setMovementStatus(movementStatus);
 
         final MovementAuthorEvaluationRequest movementAuthorEvaluationRequest = new MovementAuthorEvaluationRequest()
             .setNote(1)
