+++ src/test/java/br/com/annuum/capsicum/api/service/SaveVolunteerEvaluationServiceTest.java
@@ -6,10 +6,9 @@ import br.com.annuum.capsicum.api.domain.enums.CandidacyStatus;
 import br.com.annuum.capsicum.api.domain.enums.MovementStatus;
 import br.com.annuum.capsicum.api.exceptions.AccessControlException;
 import br.com.annuum.capsicum.api.repository.VolunteerEvaluationRepository;
+import br.com.annuum.capsicum.api.validator.SaveEvaluationValidator;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.EnumSource;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Mockito;
@@ -33,6 +32,9 @@ class SaveVolunteerEvaluationServiceTest {
 
     @Mock
     private FindCandidacyByIdService findCandidacyByIdService;
+
+    @Mock
+    private SaveEvaluationValidator saveEvaluationValidator;
 
     @Test
     public void mustSaveNewUserVolunteerEvaluation_withSuccess() {
@@ -69,6 +71,7 @@ class SaveVolunteerEvaluationServiceTest {
             .thenReturn(movement);
         Mockito.when(volunteerEvaluationRepository.save(expectedVolunteerEvaluation))
             .thenReturn(expectedVolunteerEvaluation);
+        Mockito.doNothing().when(saveEvaluationValidator).validate(movement, candidacy);
 
         //Act
         final VolunteerEvaluation returnedVolunteerEvaluation =
@@ -110,133 +113,5 @@ class SaveVolunteerEvaluationServiceTest {
 
         assertThrows(AccessControlException.class, () -> saveVolunteerEvaluationService.save(idUserEvaluator, volunteerEvaluationRequest));
     }
-
-    @Test
-    public void mustReturnExceptionWhenUserEvaluatedHasCandidacyStatusCandidate() {
-        // Arrange
-        Candidacy candidacy = new Candidacy()
-            .setUserCandidate(Mockito.mock(UserVolunteer.class))
-            .setCandidacyStatusControl(new CandidacyStatusControl());
-
-        final UserGroup userEvaluator = new UserGroup();
-        userEvaluator.setId(1L);
-
-        final Movement movement = new Movement()
-            .setUserAuthor(userEvaluator)
-            .setMovementStatus(MovementStatus.CONCLUDE);
-
-        final VolunteerEvaluationRequest volunteerEvaluationRequest = new VolunteerEvaluationRequest()
-            .setNote(1)
-            .setFeedback("someFeedBack")
-            .setIdCandidacy(1L);
-
-        final Long idUserEvaluator = 1L;
-
-        Mockito.when(findCandidacyByIdService.find(volunteerEvaluationRequest.getIdCandidacy()))
-            .thenReturn(candidacy);
-        Mockito.when(findMovimentByNeedService.find(candidacy.getNeed()))
-            .thenReturn(movement);
-
-        assertThrows(AccessControlException.class, () -> saveVolunteerEvaluationService.save(idUserEvaluator, volunteerEvaluationRequest));
-    }
-
-    @ParameterizedTest
-    @EnumSource(value = CandidacyStatus.class, names = {"APPROVED", "DECLINED", "REJECTED"})
-    public void mustReturnExceptionWhenUserEvaluatedHasCandidacyStatusApprovedOrDeclinedOrRejected(CandidacyStatus candidacyStatus) {
-        // Arrange
-        Candidacy candidacy = new Candidacy()
-            .setUserCandidate(Mockito.mock(UserVolunteer.class))
-            .setCandidacyStatusControl(new CandidacyStatusControl());
-
-        candidacy.getCandidacyStatusControl().setStatusEnum(candidacyStatus);
-
-        final UserGroup userEvaluator = new UserGroup();
-        userEvaluator.setId(1L);
-
-        final Movement movement = new Movement()
-            .setUserAuthor(userEvaluator)
-            .setMovementStatus(MovementStatus.CONCLUDE);
-
-        final VolunteerEvaluationRequest volunteerEvaluationRequest = new VolunteerEvaluationRequest()
-            .setNote(1)
-            .setFeedback("someFeedBack")
-            .setIdCandidacy(1L);
-
-        final Long idUserEvaluator = 1L;
-
-        Mockito.when(findCandidacyByIdService.find(volunteerEvaluationRequest.getIdCandidacy()))
-            .thenReturn(candidacy);
-        Mockito.when(findMovimentByNeedService.find(candidacy.getNeed()))
-            .thenReturn(movement);
-
-        assertThrows(AccessControlException.class, () -> saveVolunteerEvaluationService.save(idUserEvaluator, volunteerEvaluationRequest));
-    }
-
-
-    @Test
-    public void mustReturnExceptionWhenUserEvaluatedHasCandidacyStatusAbsent() {
-        // Arrange
-        Candidacy candidacy = new Candidacy()
-            .setUserCandidate(Mockito.mock(UserVolunteer.class))
-            .setCandidacyStatusControl(new CandidacyStatusControl());
-
-        candidacy.getCandidacyStatusControl().setStatusEnum(CandidacyStatus.APPROVED);
-        candidacy.getCandidacyStatusControl().setStatusEnum(CandidacyStatus.ABSENT);
-
-        final UserGroup userEvaluator = new UserGroup();
-        userEvaluator.setId(1L);
-
-        final Movement movement = new Movement()
-            .setUserAuthor(userEvaluator)
-            .setMovementStatus(MovementStatus.CONCLUDE);
-
-        final VolunteerEvaluationRequest volunteerEvaluationRequest = new VolunteerEvaluationRequest()
-            .setNote(1)
-            .setFeedback("someFeedBack")
-            .setIdCandidacy(1L);
-
-        final Long idUserEvaluator = 1L;
-
-        Mockito.when(findCandidacyByIdService.find(volunteerEvaluationRequest.getIdCandidacy()))
-            .thenReturn(candidacy);
-        Mockito.when(findMovimentByNeedService.find(candidacy.getNeed()))
-            .thenReturn(movement);
-
-        assertThrows(AccessControlException.class, () -> saveVolunteerEvaluationService.save(idUserEvaluator, volunteerEvaluationRequest));
-    }
-
-    @ParameterizedTest
-    @EnumSource(value = MovementStatus.class, names = {"ACTIVE", "CANCELLED"})
-    public void mustReturnExceptionWhenEvaluatingUserVolunteerMovementStatusDifferentThanConclude(MovementStatus movementStatus) {
-        // Arrange
-        Candidacy candidacy = new Candidacy()
-            .setUserCandidate(Mockito.mock(UserVolunteer.class))
-            .setCandidacyStatusControl(new CandidacyStatusControl());
-
-        candidacy.getCandidacyStatusControl().setStatusEnum(CandidacyStatus.APPROVED);
-        candidacy.getCandidacyStatusControl().setStatusEnum(CandidacyStatus.PRESENT);
-
-        final UserGroup userEvaluator = new UserGroup();
-        userEvaluator.setId(1L);
-
-        final Movement movement = new Movement()
-            .setUserAuthor(userEvaluator)
-            .setMovementStatus(movementStatus);
-
-        final VolunteerEvaluationRequest volunteerEvaluationRequest = new VolunteerEvaluationRequest()
-            .setNote(1)
-            .setFeedback("someFeedBack")
-            .setIdCandidacy(1L);
-
-        final Long idUserEvaluator = 1L;
-
-        Mockito.when(findCandidacyByIdService.find(volunteerEvaluationRequest.getIdCandidacy()))
-            .thenReturn(candidacy);
-        Mockito.when(findMovimentByNeedService.find(candidacy.getNeed()))
-            .thenReturn(movement);
-
-        assertThrows(AccessControlException.class, () -> saveVolunteerEvaluationService.save(idUserEvaluator, volunteerEvaluationRequest));
-    }
-
 
 }
